pyproject.toml
[project]
name = "dailyfin"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "playwright (>=1.51.0,<2.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=5.3.2,<6.0.0)",
    "playwright-stealth (>=1.0.6,<2.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "openai (>=1.75,<2.0)"
]

[tool.poetry]
packages = [{include = "dailyfin", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
